changed to u32

+++ 2023/day1/src/main.rs
@@ -6,9 +6,9 @@ mod day1;
 
 fn main() {
     let data = fs::read_to_string("day1input.txt").unwrap();
-    let sum: u64 = data
+    let sum: u32 = data
         .lines()
-        .map(|line| Config::from_str(line).to_num() as u64)
+        .map(|line| Config::from_str(line).to_num())
         .sum();
     println!("Sum of the configs is: {sum}");
 }
